# openspec/changes/github-global/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Global API
  description: |
    GitHub Global 后端接口文档

    ## 概述
    GitHub Global 是一个帮助开源项目作者将文档自动翻译成多种语言的 SaaS 平台。
    本 API 文档详细描述了所有后端接口的规范。

    ## 认证方式
    - 所有需要用户身份的接口使用 Session Cookie 认证
    - Cookie: `session_id`, HttpOnly, Secure, SameSite=Lax
    - OAuth 流程使用 GitHub OAuth 2.0 Authorization Code Flow

    ## 限流说明
    | 层级 | 限制 | 时间窗口 |
    |------|------|---------|
    | IP | 60 次 | 1 分钟 |
    | 用户-免费 | 100 次 | 1 小时 |
    | 用户-付费 | 1000 次 | 1 小时 |
    | 仓库 | 10 次 | 1 小时 |

    ## 错误响应格式
    所有错误响应遵循以下格式：
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "错误描述",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: GitHub Global Team
    url: https://github-global.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.github-global.example.com
    description: 生产环境服务器
  - url: http://localhost:3000
    description: 本地开发服务器
...
paths:
...
  /api/auth/callback:
    get:
      summary: GitHub OAuth 回调处理
      description: |
        GitHub OAuth 授权回调处理接口。
        用户在 GitHub 授权后，GitHub 会重定向到此接口。
      tags:
        - Authentication
      operationId: handleGitHubOAuthCallback
      parameters:
        - name: code
          in: query
          description: GitHub 授权码
          required: false
          schema:
            type: string
        - name: state
          in: query
          description: CSRF 防护状态参数
          required: false
          schema:
            type: string
        - name: error
          in: query
          description: 授权错误码
          required: false
          schema:
            type: string
        - name: error_description
          in: query
          description: 错误描述
          required: false
          schema:
            type: string
      responses:
        '302':
          description: 授权成功后重定向到首页或指定页面
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "INVALID_PARAMETER"
                  message: "缺少必要的参数"
...
  /api/user/profile:
    get:
      summary: 获取用户信息
      description: 获取当前登录用户的详细信息
      tags:
        - User
      operationId: getUserProfile
      security:
        - SessionAuth: []
      responses:
        '200':
          description: 用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
              example:
                success: true
                data:
                  id: 1
                  githubId: 12345678
                  login: "octocat"
                  name: "Octocat"
                  email: "octocat@github.com"
                  avatarUrl: "https://avatars.githubusercontent.com/u/583231"
                  isWhitelisted: false
                  quota:
                    used: 50
                    total: 100
                    resetAt: "2024-01-02T00:00:00.000Z"
                  createdAt: "2024-01-01T00:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      summary: 更新用户设置
      description: 更新用户个人设置，包括 OpenRouter API Key 等
      tags:
        - User
      operationId: updateUserProfile
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 设置更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # 仓库管理接口
  # ============================================
  /api/repos:
    get:
      summary: 获取用户仓库列表
      description: |
        获取当前用户授权访问的 GitHub 仓库列表。
        包含分页和过滤功能。
      tags:
        - Repositories
      operationId: listRepositories
      security:
        - SessionAuth: []
      parameters:
        - name: page
          in: query
          description: 页码，从 1 开始
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: 每页数量
          required: false
          schema:
            type: integer
            default: 30
            maximum: 100
        - name: sort
          in: query
          description: 排序字段
          required: false
          schema:
            type: string
            enum: [updated, created, pushed, full_name]
            default: updated
        - name: direction
          in: query
          description: 排序方向
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: type
          in: query
          description: 仓库类型过滤
          required: false
          schema:
            type: string
            enum: [all, owner, member, public, private]
            default: all
      responses:
        '200':
          description: 仓库列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryListResponse'
              example:
                success: true
                data:
                  repos:
                    - id: 1
                      githubRepoId: 12345678
                      owner: "octocat"
                      name: "hello-world"
                      fullName: "octocat/hello-world"
                      description: "A hello world repository"
                      private: false
                      defaultBranch: "main"
                      url: "https://github.com/octocat/hello-world"
                      imported: true
                      createdAt: "2024-01-01T00:00:00.000Z"
                  pagination:
                    page: 1
                    limit: 30
                    total: 100
                    hasMore: true
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: 导入仓库
      description: 将 GitHub 仓库导入到平台进行翻译管理
      tags:
        - Repositories
      operationId: importRepository
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRepositoryRequest'
      responses:
        '201':
          description: 仓库导入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 仓库已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "REPO_ALREADY_EXISTS"
                  message: "该仓库已被导入"
...
  /api/repos/{id}/files:
    get:
      summary: 获取仓库文件树
      description: |
        获取仓库的文件树结构。
        支持懒加载，仅在展开目录时获取子目录内容。
      tags:
        - Repositories
      operationId: getRepositoryFiles
      security:
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          description: 仓库 ID
          required: true
          schema:
            type: integer
        - name: path
          in: query
          description: 目录路径，空字符串表示根目录
          required: false
          schema:
            type: string
            default: ""
        - name: ref
          in: query
          description: Git 分支或提交 SHA
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 文件列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileTreeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/translate/{taskId}:
    get:
      summary: 获取翻译任务状态
      description: 获取翻译任务的详细状态和进度
      tags:
        - Translation
      operationId: getTranslationTask
      security:
        - SessionAuth: []
      parameters:
        - name: taskId
          in: path
          description: 任务 ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationTaskDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: HttpOnly Session Cookie 认证
      # Cookie 认证不需要 Authorization header
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            message:
              type: string
              example: "错误描述"
            details:
              type: object
              additionalProperties: true

    # 健康检查
...
    UserProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: integer
            githubId:
              type: integer
            login:
              type: string
            name:
              type: string
              nullable: true
            email:
              type: string
              nullable: true
            avatarUrl:
              type: string
            isWhitelisted:
              type: boolean
            quota:
              $ref: '#/components/schemas/UserQuota'
            createdAt:
              type: string
              format: date-time
    UserQuota:
      type: object
      properties:
        used:
          type: integer
        total:
          type: integer
        resetAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        openrouterApiKey:
          type: string
          description: 用户的 OpenRouter API Key（可选）
          pattern: "^sk-or-v1-[a-zA-Z0-9]+$"
        notificationEmail:
          type: string
          format: email
          description: 通知邮箱

    # 仓库相关
    RepositoryListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            repos:
              type: array
              items:
                $ref: '#/components/schemas/RepositoryItem'
            pagination:
              $ref: '#/components/schemas/Pagination'
    RepositoryItem:
      type: object
      properties:
        id:
          type: integer
        githubRepoId:
          type: integer
        owner:
          type: string
        name:
          type: string
        fullName:
          type: string
        description:
          type: string
          nullable: true
        private:
          type: boolean
        defaultBranch:
          type: string
        url:
          type: string
          format: uri
        imported:
          type: boolean
        createdAt:
          type: string
          format: date-time
    RepositoryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/RepositoryItem'
...
    ImportRepositoryRequest:
      type: object
      required:
        - owner
        - name
      properties:
        owner:
          type: string
          description: 仓库所有者
        name:
          type: string
          description: 仓库名称
...
    FileTreeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            path:
              type: string
            entries:
              type: array
              items:
                $ref: '#/components/schemas/FileEntry'
    FileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, dir, symlink]
        size:
          type: integer
          nullable: true
        sha:
          type: string
        url:
          type: string

    # 提交相关
...
    TranslationTask:
      type: object
      properties:
        id:
          type: integer
        repositoryId:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed]
        type:
          type: string
          enum: [full, incremental]
        targetLanguages:
          type: array
          items:
            type: string
        totalFiles:
          type: integer
        processedFiles:
          type: integer
        failedFiles:
          type: integer
        progress:
          type: number
        errorMessage:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    TranslationTaskDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            task:
              $ref: '#/components/schemas/TranslationTask'
            files:
              type: array
              items:
                $ref: '#/components/schemas/TranslationFile'
    TranslationFile:
      type: object
      properties:
        path:
          type: string
        status:
          type: string
          enum: [pending, translating, completed, failed]
        originalContent:
          type: string
        translatedContent:
          type: string
          nullable: true
        language:
          type: string
        error:
          type: string
          nullable: true
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_PARAMETER"
              message: "请求参数验证失败"
              details:
                field: "email"
                issue: "无效的邮箱格式"
    Unauthorized:
      description: 未授权，需要登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "请先登录"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "资源不存在"
...